V1
Complete file upload of all scripts and modules used to create the planetfind aid on a Raspberry Pi 3 Model B+. As a note files deltat.data, deltat.preds and Leap_Second.dat are part of the Skyfield API library, de421.bsp (found at https://rhodesmill.org/skyfield/). This commit only utilises de421.bsp due to only requiring current coordinates of respective planets.

The core script for running the app is path.sh. This script is a bash script that calls location.py and planet.py. Also included is the cron job that is currently running on my pi to start the app on boot.
The keys.py script holds keys that allow the app to write to Twitter.

This application works in conjunction with the Blynk app(https://blynk.io/) for mobile devices. The virtual pin selected for the GPS stream widget in the app will need to be matched with the code here or customised bespoke.  
Similarly the Twython keys in the key.py file will have to be swapped out for the respective keys provided to the user when setting up a developer API with Twitter.